+++ IsekaiMod/Content/Classes/IsekaiProtagonist/IsekaiProtagonistClass.cs
@@ -104,7 +104,7 @@ namespace IsekaiMod.Content.Classes.IsekaiProtagonist
                     StatType.SkillUseMagicDevice
                 };
                 bp.IsDivineCaster = true;
-                bp.IsArcaneCaster = false;
+                bp.IsArcaneCaster = true;
                 bp.StartingGold = 69420;
                 bp.PrimaryColor = 9;
                 bp.SecondaryColor = 9;